URL converter added.

+++ DCMS/src/main/java/com/nima/dcms/urlconverter/URLConverter.kt
@@ -1,5 +1,5 @@
 package com.nima.dcms.urlconverter
 
 interface URLConverter {
-    fun convert(url: String): Int
+    fun convert(url: String): Long
 }
